aihub-compass-setup/03-agent-test/pyproject.toml
[project]
name = "compass-agent-test"
version = "0.1.0"
description = "Test scripts for Foundry Agents using Core42 Compass via APIM"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=2.0.0b2",
    "azure-identity>=1.26.0b1",
    "python-dotenv>=1.2.1",
]
